Full Teleop
I made a teleop program that uses all the necessary buttons in order to
make the robot perform every task that it is designed to carry out. As
soon as we have a completed robot we can test it in its entirety. As a
side note, it should be easy to adapt this code to use multiple
controllers.

The other program in this commit is one we used testing the opening and
closing of the door on the ball trough.